requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
google-api-core==1.1.2
google-auth==1.4.1
google-cloud-vision==0.31.1
googleapis-common-protos==1.5.3
grpcio==1.12.0
idna==2.6
pkg-resources==0.0.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2018.4
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
